move cmd/common_test to internal/tests/common
