+++ cmd/cbootstrapd/api_test.go
@@ -14,7 +14,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/mistifyio/lochness"
-	"github.com/mistifyio/lochness/cmd/common_test"
+	"github.com/mistifyio/lochness/internal/tests/common"
 	logx "github.com/mistifyio/mistify-logrus-ext"
 	"github.com/stretchr/testify/suite"
 )
