+++ cmd/nheartbeatd/main_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/mistifyio/lochness"
-	"github.com/mistifyio/lochness/cmd/common_test"
+	"github.com/mistifyio/lochness/internal/tests/common"
 	"github.com/pborman/uuid"
 	"github.com/stretchr/testify/suite"
 )
