+++ internal/tests/common/msg.go
